fix numpy.ndarray mypy errors

+++ bdd100k/eval/mots.py
@@ -38,7 +38,7 @@ def acc_single_video_mots(  # pylint: disable=unused-argument
         if not result:
             res_masks = gen_blank_bitmask(gt_masks.shape)
         else:
-            res_masks = np.asarray(Image.open(result))
+            res_masks = np.asarray(Image.open(result), dtype=np.uint8)
         gt_masks, gt_ids, gt_attrs, gt_cats = parse_bitmasks(gt_masks)
         pred_masks, pred_ids, pred_attrs, pred_cats = parse_bitmasks(res_masks)
         ious, iofs = bitmask_intersection_rate(gt_masks, pred_masks)
